backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
supabase==1.2.0
python-decouple==3.8